day 17 part 2, finally. off by one error with jet pattern

+++ 2022/aoc2022.day17/domain/HorizontalLine.cs
@@ -2,7 +2,7 @@
 {
     public class HorizontalLine : Shape
     {
-        public override bool IsBlockedBelow(HashSet<int>[] heights, int bottomLeftIndex, int bottomLeftHeight)
+        public override bool IsBlockedBelow(HashSet<long>[] heights, int bottomLeftIndex, long bottomLeftHeight)
         {
             return heights[bottomLeftIndex].Contains(bottomLeftHeight - 1) ||
                    heights[bottomLeftIndex + 1].Contains(bottomLeftHeight - 1) ||
@@ -10,20 +10,20 @@
                    heights[bottomLeftIndex + 3].Contains(bottomLeftHeight - 1);
         }
 
-        public override bool IsBlockedToTheLeft(HashSet<int>[] heights, int bottomLeftIndex, int bottomLeftHeight)
+        public override bool IsBlockedToTheLeft(HashSet<long>[] heights, int bottomLeftIndex, long bottomLeftHeight)
         {
             return bottomLeftIndex == 0 || 
                    heights[bottomLeftIndex - 1].Contains(bottomLeftHeight);
         }
 
-        public override bool IsBlockedToTheRight(HashSet<int>[] heights, int bottomLeftIndex, int bottomLeftHeight)
+        public override bool IsBlockedToTheRight(HashSet<long>[] heights, int bottomLeftIndex, long bottomLeftHeight)
         {
             var rightWallIndex = heights.Length - 1;
             return bottomLeftIndex == rightWallIndex - 3 ||
                    heights[bottomLeftIndex + 4].Contains(bottomLeftHeight);
         }
 
-        public override void UpdateHeightsAfterComingToRest(ref HashSet<int>[] heights, int bottomLeftIndex, int bottomLeftHeight)
+        public override void UpdateHeightsAfterComingToRest(ref HashSet<long>[] heights, int bottomLeftIndex, long bottomLeftHeight)
         {            
             // ......
             // .****.
