+++ 2022/aoc2022.day17/domain/MirroredL.cs
@@ -2,14 +2,14 @@
 {
     public class MirroredL : Shape
     {
-        public override bool IsBlockedBelow(HashSet<int>[] heights, int bottomLeftIndex, int bottomLeftHeight)
+        public override bool IsBlockedBelow(HashSet<long>[] heights, int bottomLeftIndex, long bottomLeftHeight)
         {
             return heights[bottomLeftIndex].Contains(bottomLeftHeight - 1) ||
                    heights[bottomLeftIndex + 1].Contains(bottomLeftHeight - 1) ||
                    heights[bottomLeftIndex + 2].Contains(bottomLeftHeight - 1);
         }
 
-        public override bool IsBlockedToTheLeft(HashSet<int>[] heights, int bottomLeftIndex, int bottomLeftHeight)
+        public override bool IsBlockedToTheLeft(HashSet<long>[] heights, int bottomLeftIndex, long bottomLeftHeight)
         {
             return bottomLeftIndex == 0 ||
                    heights[bottomLeftIndex - 1].Contains(bottomLeftHeight) ||
@@ -17,7 +17,7 @@
                    heights[bottomLeftIndex + 1].Contains(bottomLeftHeight + 2);
         }
 
-        public override bool IsBlockedToTheRight(HashSet<int>[] heights, int bottomLeftIndex, int bottomLeftHeight)
+        public override bool IsBlockedToTheRight(HashSet<long>[] heights, int bottomLeftIndex, long bottomLeftHeight)
         {
             var rightWallIndex = heights.Length - 1;
             // return bottomLeftIndex == rightWallIndex - 2 ||
@@ -28,7 +28,7 @@
                    heights[bottomLeftIndex + 3].Contains(bottomLeftHeight + 2);
         }
 
-        public override void UpdateHeightsAfterComingToRest(ref HashSet<int>[] heights, int bottomLeftIndex, int bottomLeftHeight)
+        public override void UpdateHeightsAfterComingToRest(ref HashSet<long>[] heights, int bottomLeftIndex, long bottomLeftHeight)
         {
             // .....
             // ...*.
