+++ 2022/aoc2022.day17/domain/Shape.cs
@@ -6,13 +6,13 @@ namespace aoc2022.day17.domain
     {
         public Shape() { }
 
-        public abstract bool IsBlockedBelow(HashSet<int>[] heights, int bottomLeftIndex, int bottomLeftHeight);
+        public abstract bool IsBlockedBelow(HashSet<long>[] heights, int bottomLeftIndex, long bottomLeftHeight);
 
-        public abstract bool IsBlockedToTheLeft(HashSet<int>[] heights, int bottomLeftIndex, int bottomLeftHeight);
+        public abstract bool IsBlockedToTheLeft(HashSet<long>[] heights, int bottomLeftIndex, long bottomLeftHeight);
 
-        public abstract bool IsBlockedToTheRight(HashSet<int>[] heights, int bottomLeftIndex, int bottomLeftHeight);
+        public abstract bool IsBlockedToTheRight(HashSet<long>[] heights, int bottomLeftIndex, long bottomLeftHeight);
 
-        public abstract void UpdateHeightsAfterComingToRest(ref HashSet<int>[] heights, int bottomLeftIndex, int bottomLeftHeight);
+        public abstract void UpdateHeightsAfterComingToRest(ref HashSet<long>[] heights, int bottomLeftIndex, long bottomLeftHeight);
 
         public static Shape MakeShape(RockShape rockShape)
         {
