+++ 2022/aoc2022.day17/domain/Square.cs
@@ -2,13 +2,13 @@
 {
     public class Square : Shape
     {
-        public override bool IsBlockedBelow(HashSet<int>[] heights, int bottomLeftIndex, int bottomLeftHeight)
+        public override bool IsBlockedBelow(HashSet<long>[] heights, int bottomLeftIndex, long bottomLeftHeight)
         {
             return heights[bottomLeftIndex].Contains(bottomLeftHeight - 1) ||
                    heights[bottomLeftIndex + 1].Contains(bottomLeftHeight - 1);
         }
 
-        public override bool IsBlockedToTheLeft(HashSet<int>[] heights, int bottomLeftIndex, int bottomLeftHeight)
+        public override bool IsBlockedToTheLeft(HashSet<long>[] heights, int bottomLeftIndex, long bottomLeftHeight)
         {
             //     return bottomLeftIndex == 0 ||
             //            heights[bottomLeftIndex - 1].Overlaps(Enumerable.Range(bottomLeftHeight, 2));
@@ -17,7 +17,7 @@
                    heights[bottomLeftIndex - 1].Contains(bottomLeftHeight + 1);
         }
 
-        public override bool IsBlockedToTheRight(HashSet<int>[] heights, int bottomLeftIndex, int bottomLeftHeight)
+        public override bool IsBlockedToTheRight(HashSet<long>[] heights, int bottomLeftIndex, long bottomLeftHeight)
         {
             var rightWallIndex = heights.Length - 1;
             // return bottomLeftIndex == rightWallIndex - 1 ||
@@ -27,7 +27,7 @@
                    heights[bottomLeftIndex + 2].Contains(bottomLeftHeight + 1);
         }
 
-        public override void UpdateHeightsAfterComingToRest(ref HashSet<int>[] heights, int bottomLeftIndex, int bottomLeftHeight)
+        public override void UpdateHeightsAfterComingToRest(ref HashSet<long>[] heights, int bottomLeftIndex, long bottomLeftHeight)
         {
             // ....
             // .**.
